move record.go content to layout.go
